services/agent/requirements.txt
# LiveKit Agent Requirements
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-cartesia
python-dotenv
